+++ src/autofonce_lib/filter.ml
@@ -26,6 +26,7 @@ let tests_ids = ref ( [] : ( int * int ) list )
 let tests_keywords = ref ( [] : string list )
 let tests_nokeywords = ref ( [] : string list )
 let only_failed = ref false
+let all_keywords = ref false
 
 let select_tests ?state select_test suite =
   let ntests = suite.suite_ntests in
@@ -52,6 +53,7 @@ let select_tests ?state select_test suite =
       | [] -> ()
       | ids ->
           List.iter (fun s ->
+              let s = String.lowercase_ascii s in
               let len = String.length s in
               if len>0 && s.[0] = '-' then
                 no_set := StringSet.add ( String.sub s 1 (len-1) ) !no_set
@@ -72,8 +74,11 @@ let select_tests ?state select_test suite =
       && t.test_id <= !exec_before
       && (all_tests
           || id_set. (t.test_id)
-          || List.exists (fun k -> StringSet.mem k keyword_set)
-            t.test_keywords
+          ||
+          (if !all_keywords then
+             List.for_all (fun k ->  StringSet.mem k keyword_set) t.test_keywords
+           else
+             List.exists (fun k ->  StringSet.mem k keyword_set) t.test_keywords)
          )
       && not (
           List.exists (fun k -> StringSet.mem k nokeyword_set)
@@ -203,6 +208,9 @@ let args = [
       clean_tests_dir := false
     ),
   EZCMD.info "Run only previously failed tests (among selected tests)";
+
+  [ "match-all" ], Arg.Set all_keywords,
+  EZCMD.info "Run tests matching all keywords instead of only one";
 
   [], Arg.Anons (fun list ->
       match list with